DB: enforce foreign keys + sales_items.quantity int->double (schema v6)
Two pre-feature schema fixes from the database review, done while the tables
are still small:

- Enforce foreign keys per-connection via PRAGMA foreign_keys = ON in
  MigrationStrategy.beforeOpen (runs after migrations, where table rebuilds
  need FKs off). No FK constraints are declared yet, so it's a safe no-op
  guard today; any references(...) added by a future feature (purchases,
  debts, stock movements) will now actually be enforced.

- sales_items.quantity int -> double, matching products.quantity so weight/
  fractional sales can be recorded and future *_items tables share the type.
  SQLite can't change a column type in place, so the v5->v6 migration rebuilds
  the table via alterTable(TableMigration(salesItems)) and re-runs
  _createIndexes() (the rebuild drops the table's indexes). Ripple: InvoiceItem
  .quantity -> double, _onConfirmSale converts the cart's whole-unit int with
  .toDouble(), and history formats whole numbers as "2x" not "2.0x". The cart
  UI stays int.

Regenerated Drift code; flutter analyze clean; smoke tests 6/6.

+++ lib/features/billing/domain/entities/invoice_item.dart
@@ -7,7 +7,7 @@ class InvoiceItem extends Equatable {
   final String productName;
   final double price;
   final double cost; // snapshot of product cost at sale time (profit reports)
-  final int quantity;
+  final double quantity; // double so weight/fractional sales can be recorded
 
   const InvoiceItem({
     this.id,
